requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.1
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.1
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
configobj==5.0.6
constantly==15.1.0
cryptography==3.0
daphne==3.0.1
Django==3.1
django-admin-relation-links==0.2.5
django-advanced-filters==1.3.0
django-appconf==1.0.4
django-braces==1.14.0
django-colorfield==0.4.1
django-ipware==3.0.2
django-jet==1.0.8
django-json-widget==1.1.1
django-loginas==0.3.9
django-modeladmin-reorder==0.3.1
django-nested-admin==3.3.3
django-pipeline==2.0.6
django-recaptcha==2.0.6
django-recaptcha3==0.4.0
django-redis==4.12.1
django-select2==7.7.1
django-tinymce==3.2.0
djangorestframework==3.12.2
dropbox==11.0.0
future==0.18.2
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
httpie==2.3.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.2
msgpack-python==0.5.6
packaging==20.9
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.0.1
ply==3.11
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
PySocks==1.7.1
python-bidi==0.4.2
python-dotenv==0.17.1
python-gnupg==0.4.6
python-ldap==3.3.1
python-monkey-business==1.0.0
pytz==2020.1
redis==3.5.3
reportlab==3.5.57
requests==2.25.1
requests-toolbelt==0.9.1
secrets==1.0.2
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
stone==3.2.1
systematic==4.8.7
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
webencodings==0.5.1
websockets==8.1
xhtml2pdf==0.2.5
zope.interface==5.1.0
